requirements.txt
streamlit
pandas
matplotlib
reportlab
google-genai
google-adk
python-dotenv
